Adjust jump-to-bottom behavior
User's chat will now only jump to the bottom on their own posted messages, so they won't be repeatedly brought
to the bottom while trying to read through old messages in an active channel. Wrapped the dispatched version of
receiveMessage in a Promise to ease timing of scroll as well as cleanly passing relevant info. Made adjustments
to backend to accomodate for reduced componentDidUpdate checks, then adjusted reducers in kind.