Fix message alert for DMs
Alerting the user to new messages while scrolled up from the bottom of the chat was functioning for channels
but not for DMs. Discovered that this was due to timing issues in the Channel page (the 'bottom' element
did not exist during first render, and other timing issues occured. Moved and adjusted IntersectionObserver
logic, cleaning up creation of observers as well.