Complete functionality of New DMs
DMs can now be created on the fly! Search for other users by email or
username. New DM not created until user submits the message, at which point
the controller will either create a new DM (and broadcast a message out
via the MembershipsChannel connections to the relevant users) or add
the message to the existing DM if there is one.

Some style tweaks still necessary for searchbar.