Clear input and focus style switch for names
Now the input in the user search for a new DM will clear when changing focus to anything
other than it and the dropdown. In addition, though style fixes are still necessary, the
user clicking into the textarea in the message form will change the searchbar to instead
be a display of the name values until the searchbar area is focused again.