Add ActionCable current_user and broadcast fix
Implemented current_user in action cable connection for memberships
to use the session cookie rather than passing the id through as a
parameter to creating a new Memberships subscription. Cleaning
implementation of subscription also led to discovery of inefficiently
recreating the same subscription many times. No longer an issue
as subscribing only runs when the App is mounted.